refactor: CardList Component 분리
