+++ src/pages/CardList/components/index.ts
@@ -1,0 +1,1 @@
+export { default as Cards } from './Cards';
